force main window to display on notification click
A notification click propagates as a click to the newly received message.
However, the click did not force the Inboxer window to be displayed. If
the window was minimized or hidden, the user may not notice the effects of
the click until next time the Inboxer window is shown.

With this chance, the Inboxer window is displayed and the click is
propagated, so the user can see its effect.