select folder by loading correct URL
First try to locate folder link on the left. If this fails
try to load the correct URL: "mail.google.com/mail/#folder".

This makes it more robust and allows to always go back to Inbox
even when the window has navigated somewhere it shouldn't have.
For example, on clicking "Add Account" we navigate to a new
login page from which we cannot come back.