notify when downloading a file
Clicking on the download button of an attachment causes the
file to be silently downloaded to the Downloads folder.
The user does not receive any feedback and is left wondering if the
download worked at all (possibly related to issue #57).

This change notifies the user of the successful download and the
location of the file.

NOTE: while testing this commit I've sometimes (very rarely) seen a
weird behavior where clicking a download button fails silently with
this error in the JS console:
Refused to display 'url' in a frame because it set 'X-Frame-Options' to 'SAMEORIGIN'.

This issue is beyond our control (caused by gmail?). At least now the
error can be noticed by the lack of notification.