display migration dialog
Display a dialog on startup to inform the user that this version has been
migrated to Gmail. The dialog explains about Google's decision to
discontinue Inbox and details which versions still work with Inbox and
which with Gmail.

We add a new config setting 'displayMigrationInfo' to allow the user
to forever dismiss the dialog.